requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.24.0
peft>=0.7.0
bitsandbytes>=0.41.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
rouge-score>=0.1.2
nltk>=3.8.0
pyyaml>=6.0
tqdm>=4.66.0
jinja2>=3.1.0
pytest>=7.4.0
datasets>=2.14.0
